2024/02/15

+++ src/main/java/com/greengram/greengram4/feed/model/FeedPicsInsDto.java
@@ -7,6 +7,6 @@ import java.util.List;
 
 @Data
 public class FeedPicsInsDto {
-    private int ifeed;
+    private Long ifeed;
     private List<String> pics = new ArrayList<>();
 }
