+++ src/main/java/com/greengram/greengram4/user/model/UserFollowDto.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class UserFollowDto {
-    private int fromIuser;
-    private int toIuser;
+    private Long fromIuser;
+    private Long toIuser;
 }
